# backend/docs/swagger.yaml
basePath: /
definitions:
...
  api.DeletedResponse:
    properties:
      deleted:
        type: boolean
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.IDResponse:
    properties:
      id:
        type: integer
    type: object
...
host: localhost:8011
info:
  contact:
    email: support@callified.ai
    name: Callified Support
  description: AI-powered outbound calling, WhatsApp CRM & receptionist platform.
  termsOfService: http://callified.ai/terms
  title: Callified Backend API
  version: "1.0"
paths:
...
  /api/wa/session:
    get:
      description: Returns WaSender session list for the org's Personal Access Token.
        Requires Admin role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List WaSender sessions
      tags:
      - whatsapp
  /api/wa/session/{id}/connect:
    post:
      description: Initiates connection and returns the initial QR code for scanning.
        Requires Admin role.
      parameters:
      - description: WaSender session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Connect WaSender session
      tags:
      - whatsapp
  /api/wa/session/{id}/disconnect:
    post:
      description: Drops the WhatsApp link and resets session to NEED_SCAN state.
        Requires Admin role.
      parameters:
      - description: WaSender session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disconnect WaSender session
      tags:
      - whatsapp
  /api/wa/session/{id}/qr:
    get:
      description: Fetches a fresh QR code for an initialized session (used for 40s
        auto-refresh). Requires Admin role.
      parameters:
      - description: WaSender session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get WaSender session QR
      tags:
      - whatsapp
...
  /api/webhooks:
    get:
      description: Returns all configured outbound webhooks for the org. Requires
        Admin role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List webhooks
      tags:
      - webhooks
    post:
      consumes:
      - application/json
      description: Registers a new outbound webhook endpoint. Requires Admin role.
      parameters:
      - description: Webhook config
        in: body
        name: body
        required: true
        schema:
          properties:
            event:
              type: string
            secret_key:
              type: string
            url:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create webhook
      tags:
      - webhooks
  /api/webhooks/{id}:
    delete:
      description: Removes an outbound webhook. Requires Admin role.
      parameters:
      - description: Webhook ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DeletedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete webhook
      tags:
      - webhooks
  /api/webhooks/{id}/logs:
    get:
      description: Returns the last 50 delivery attempts for a webhook. Requires Admin
        role.
      parameters:
      - description: Webhook ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get webhook logs
      tags:
      - webhooks
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: **Bearer {JWT token}**. Obtain a token from POST /api/auth/login
      or POST /api/auth/signup.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"